requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.6.0
PyJWT>=2.8.0
